Spec for stack lifecycle plugpoint
A cloud provider may have a need for custom code to examine stack requests
prior to performing the operations to create or update a stack.
Some providers may also need code to run after operations on
a stack complete. A mechanism is proposed whereby providers may easily add
pre-operation calls from heat to their own code, which is called prior to
performing the stack work, and post-operation calls, which are made after
a stack operation completes or fails.

blueprint stack-lifecycle-plugpoint

Change-Id: Ide4517f55267a1978ffffb264acd8a8128f38835