Reorg AutoScalingGroup code
The current class hierarchy of autoscaling groups in Heat is hindering
further improvements to both AWS version and OS version.  On one hand,
we don't want OS resources to extend AWS resources; on the other hand,
we have a lot proposals to enhance the OS version.  Also, the classes
should be split and relocated into different subdirectories.

Change-Id: Idedc141b9f63e4ed20b54bd297d2f737fe1249d1