Add template capabilities
Outlines a plan for potentially enhancing HOT to support an optional
capabilities annotation, which is similar in concept to the
TOSCA substitution_mapping definition, although we only consider the
matching of capabilities in this spec (not specific resource types,
although a method for doing that is discussed).

The aim here is to enable easier composability, but providing built-in
support for compatible matching templates to resource_registry
mappings.

Change-Id: Ie1b3c564b38f929d587af7074c1327c495846e57
Co-Authored-By: Jay Dobies <jason.dobies@redhat.com>