New Resource Type OS::Cinder::Quota
This creates new resource type whose intended use case is for admin-only use
and specifically to manage the cinder quota. This would be an enhancement to
avoid manual, post-deployment modifications of the project's cinder quota.
We want to enable the admin to fully orchestrate the project by using heat
without the need to make manual updates afterwards.

Change-Id: I119391118a4d83b02fe0f5773665136de163f3e8