New Resource Type OS::Nova::Quota
This creates new resource type whose intended use case is for
admin-only use and specifically to manage the Compute service
quota. This would be an enhancement to avoid manual, post-
deployment modifications of the project's (and user's) Compute
service quota. This will enable the admin to fully orchestrate
the project by using heat without the need to make manual
updates afterwards.

Change-Id: I78cb376d60a826d21c3bad9657bda11dcc51c3f1